chore: splitting utils
